+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.6.2] - Unreleased
+
+### Added
+
+- Added the attribute `protected` to the resource `neon_branch` to provision protected branches.
+
+### Fixed
+
+- [[#108](https://github.com/kislerdm/terraform-provider-neon/issues/108)] Fixed import of the resource `neon_role`.
+- Fixed mutability for the `brach` attribute of the `neon_project` resource.
+
+### Changed
+
+- Updated dependencies:
+  - Neon Go SDK: [v0.6.1](https://github.com/kislerdm/neon-sdk-go/compare/v0.5.0...v0.6.1)
+
 ## [v0.6.1] - 2024-09-28
 
 ### Added